feat: Refactor HTTP response handling in server
The code changes in `main.js` refactor the HTTP response handling in the server. It introduces a `response` object that encapsulates the different types of HTTP responses, such as `OK` and `E404`. It also adds a `send` method to the `response` object, which allows sending custom data in the response body. The server now uses the `response` object to send appropriate responses based on the requested URI.